% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_statsguru_results.R
\name{load_statsguru_results}
\alias{load_statsguru_results}
\title{Load Cricinfo Statsguru Results Tables into R}
\usage{
load_statsguru_results(url)
}
\arguments{
\item{url}{The url of the results page}
}
\description{
This function takes a statsguru results page url and returns every page of results as a single data frame
}
